+++ package.json
@@ -24,6 +24,7 @@
     "framer-motion": "^11.2.12",
     "latest": "^0.2.0",
     "lodash": "^4.17.21",
+    "loglevel": "^1.9.1",
     "next": "14.2.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -33,6 +34,8 @@
     "react-toastify": "^10.0.5",
     "redux": "^5.0.1",
     "swiper": "^11.1.4",
+    "uuid": "^10.0.0",
+    "winston": "^3.13.1",
     "yup": "^1.4.0"
   },
   "devDependencies": {
@@ -40,6 +43,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/uuid": "^10.0.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
     "postcss": "^8",
