--- src/interfaces/UserInterface.ts
@@ -1,24 +1,0 @@
-export interface User {
-  email: string;
-  name: string;
-  nickname: string;
-  sid: string; // session id
-}
-
-export interface UserState {
-  isAuthenticated: boolean;
-  user: User | null;
-}
-
-export interface Attendee {
-  eventID: string;
-  name: string;
-  email: string;
-  phone: string;
-  gender: string;
-  id: string;
-}
-
-export interface AttendeesData {
-  attendees: Attendee[];
-}
